+++ src/main/java/com/example/demo/ProductoServicioImpl.java
@@ -15,7 +15,7 @@ public class ProductoServicioImpl implements ProductoServicio{
     }
 
     @Override
-    public Productos Buscar(int id) {
+    public Productos Buscar(long id) {
         for(Productos busqueda: Lista){
             if(busqueda.getId()==id)
                 return busqueda;
@@ -47,7 +47,7 @@ public class ProductoServicioImpl implements ProductoServicio{
     }
 
     @Override
-    public void Borrar(int id) {
+    public void Borrar(long id) {
         for(Productos borrar: Lista){
             if(borrar.getId()==id){
                 Lista.remove(borrar);
